+++ HISTORY.rst
@@ -10,6 +10,7 @@ development (in progress)
 * Mention `django-bootstrap4 <https://github.com/zostera/django-bootstrap4/>`_ in README
 * Rewrite `tox` test matrix to focus on Django releases rather than Python versions
 * Add tests for Django master branch (>= 2)
+* Add `label` override for `{% bootstrap_field %}`
 
 
 9.0.0 (2017-07-11)